Full 1.20.4 Support (#191)
* Update gradle.properties to iterate mod_version

Updated mod_version to ensure compatibility with FabricTailor is retained

* Update Taterzens.java to init lang as the correct datatype

Initialised an empty JsonObject to ensure the mod doesn't trip over lang being null.  Has the odd side effect of allowing the GUI to start working.

* Update LanguageUtil.java Language file path

Updated the path to the language files within the .jar.  Note that it loads from within it, NOT from the filepath in the config folder.

* Update TextUtil.java Commenting out SERVER_TRANSLATION

Commented out SEVER_TRANSLATIONS reference because it was blocking any attempt to read from the language.json files in toto.  This may be removed in future as it doesn't seem to apply anywhere else in the codebase, and was returning nothing that could be used by the functions seeking translation.

* Update EditorGUI.java Generate UUID to prevent GameProfile null crash

Generated temporary UUID to keep GameProfile from crashing the server when it tries to generate a result using a null uuid entry.

* Update Taterzens.java Syntax Error Fix

Fix syntax error for defining JsonObject

* Update TaterzensClient.java

This just seems to be dead code.  It throws an error at build when not commented out, but doesn't seem to affect mod function.  Most of the imports can be removed, too.

* Update TypeCommand.java

Introduced the concept of modEntity, which is a part of an NPC's definition, and which maps directly to the Entity Type in CAPS.

* Update NPCData.java

Added the hashmaps for the entity list and player entity so that we can track and manipulate them with ease when setting TYPE.

* Update TaterzenNPC.java

We make the necessary changes to allow player TYPE to be edited and for it to persist across world / game loads.

* Update TaterzenNPC.java

Loose bracket...