Initial 1.21 support (#192)
* Migrate to polymer

* npc selections glow again

* skin layers working again

* set skin layers to all on by default

* Fix #137 (rewrite/polymer) (#181)

* fix #137 (init): speak distance, requires some testing

* fix: removed refmap field in common.mixins.json, fixed crash
- Replacement was not anywhere in build setup
- Loom automatically adds the right refmap anyway, no need to explicitly define it

* removed commented code (lol)

* Update gradle.properties to iterate mod_version

Updated mod_version to ensure compatibility with FabricTailor is retained

* Update Taterzens.java to init lang as the correct datatype

Initialised an empty JsonObject to ensure the mod doesn't trip over lang being null.  Has the odd side effect of allowing the GUI to start working.

* Update LanguageUtil.java Language file path

Updated the path to the language files within the .jar.  Note that it loads from within it, NOT from the filepath in the config folder.

* Update TextUtil.java Commenting out SERVER_TRANSLATION

Commented out SEVER_TRANSLATIONS reference because it was blocking any attempt to read from the language.json files in toto.  This may be removed in future as it doesn't seem to apply anywhere else in the codebase, and was returning nothing that could be used by the functions seeking translation.

* Update EditorGUI.java Generate UUID to prevent GameProfile null crash

Generated temporary UUID to keep GameProfile from crashing the server when it tries to generate a result using a null uuid entry.

* Update Taterzens.java Syntax Error Fix

Fix syntax error for defining JsonObject

* Update TaterzensClient.java

This just seems to be dead code.  It throws an error at build when not commented out, but doesn't seem to affect mod function.  Most of the imports can be removed, too.

* Update TypeCommand.java

Introduced the concept of modEntity, which is a part of an NPC's definition, and which maps directly to the Entity Type in CAPS.

* Update NPCData.java

Added the hashmaps for the entity list and player entity so that we can track and manipulate them with ease when setting TYPE.

* Update TaterzenNPC.java

We make the necessary changes to allow player TYPE to be edited and for it to persist across world / game loads.

* Update TaterzenNPC.java

Loose bracket...

* Update build.gradle

Changes to exclude disguiselib (polymer provides that function), servertranslations (unsure of 1.21 version support, but doesn't impact most mod function), and updated Java build versions and targets inline with the requirements for 1.21

* Update gradle.properties

Updated the Fabric Properties to 1.21 requirements.  Updated the mod_version, and dependencies that support 1.21

* Update TextUtil.java

toJson and fromJson now need a provider as the final argument when invoked.  Currently nulled, and seems to work fine.  May require more investigation later.

* Update FabricPlatform.java

Commented out the openEditorGui code (and import) since we've disabled the GUI for now as part of the 1.21 update.

The issues is tied to ItemStacks and whether we'll stick to the official mappings or Yarn going forward.

* Update Taterzens.java

Changes to ResourceLocation introduced in Minecraft 1.21

* Update NpcCommand.java

Commented out the teleport command for now.  TODO work to correct its implementation with Minecraft 1.21 changes.

* Update TaterzensCommand.java

Updated the ResourceLocation code to match Minecraft 1.21 support

* Update TeleportCommand.java

Temporarily disables the teleport command while working out the 1.21 Minecraft changes.

* Update BehaviourCommand.java

Updated to address the ResourceLocation changes in 1.21

* Update MovementCommand.java

Updated to address the ResourceLocation changes for 1.21

* Update TaterzensCommand.java

Error when running prior diff.  Corrected the corrected ResourceLocation fix.

* Update BehaviourCommand.java

Corrected error from prior diff.  Related to the ResourceLocation fix.

* Update ProfessionsCommand.java

Updated the ResourceLocation implementation in accordance with 1.21

* Update SoundCommand.java

Updated ResourceLocation usage in accordance with the 1.21 ResourceLocation implementation.

* Update TypeCommand.java

Commenting out the DisguiseLib support for now, as Polymer provides the entity type modification interface.

* Update CommandsCommand.java

Updated the ResourceLocation code in accordance with the 1.21 ResourceLocation changes.

* Update ServerGamePacketListenerImplMixin_MsgEditor.java

Updated fromJson implementation with changes to Serializer with Minecraft 1.21

* Update TaterzenNPC.java

Significant changes throughout:

- setMaxUpStep is removed from 1.20.5
- ResourceLocation changes to meet 1.21
- dropCustomDeathLoot issues to be resolved
- isLeashed issues to be resolved
- Projectile related issues to be resolved
- Issues with data.removeIf implementation to be resolved
-

* Update BungeeCommand.java

Updated ResourceLocation code to 1.21 implementation.  Commented out ClientBoundCustomPayloadPacket code as it still needs work to meet 1.21.

Query whether Bungee support is still required in the mod.

* Update ScarpetProfession.java

Updated the ResourceLocation code to meet 1.21

* Update FabricPlatform.java

Fixed spelling mistake.  Main update was to temporarily comment out the EditorGUI functions.

* Update MessagesEditGUI.java

Commented out issues with itemStack implementation that haven't been resolved to the 1.21 implementation yet.  As such, the GUIs have been disabled for now.

* Update ListItemsGUI.java

Commented out issues with itemStack implementation that haven't been resolved to the 1.21 implementation yet.  As such, the GUIs have been disabled for now.

* Update EditorGUI.java

Commented out issues with itemStack implementation that haven't been resolved to the 1.21 implementation yet.  As such, the GUIs have been disabled for now.

---------

Co-authored-by: samolego <34912839+samolego@users.noreply.github.com>
Co-authored-by: MagicSweet <49122140+MagicSweet-dev@users.noreply.github.com>